requirements.txt
fastapi
uvicorn
python-multipart
groq
pytesseract
pillow
python-dotenv
pandas
openpyxl
